Makes it easier for other plugins to reuse template testing functions
